# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A secure file storage REST API with S3 and CloudFront",
        "title": "File Vault API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/mohamed8eo/file-vault",
            "email": "support@filevault.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": ["http", "https"],
    "paths": {
...
        "/upload": {
            "post": {
                "tags": ["Files"],
                "summary": "Upload file",
                "description": "Upload any file type (max 50MB)",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "security": [{"ApiKeyAuth": []}],
                "parameters": [{
                    "name": "file",
                    "in": "formData",
                    "required": true,
                    "type": "file",
                    "description": "File to upload"
                }],
                "responses": {
                    "200": {"description": "File uploaded successfully"},
                    "400": {"description": "Invalid file"},
                    "413": {"description": "File too large"}
                }
            }
        },
        "/upload/image": {
            "post": {
                "tags": ["Files"],
                "summary": "Upload image",
                "description": "Upload an image file (jpg, png, gif, webp, svg - max 10MB)",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "security": [{"ApiKeyAuth": []}],
                "parameters": [{
                    "name": "file",
                    "in": "formData",
                    "required": true,
                    "type": "file",
                    "description": "Image file"
                }],
                "responses": {
                    "200": {"description": "Image uploaded successfully"},
                    "400": {"description": "Invalid file type"},
                    "413": {"description": "File too large"}
                }
            }
        },
        "/upload/video": {
            "post": {
                "tags": ["Files"],
                "summary": "Upload video",
                "description": "Upload a video file (mp4, webm, mov, avi, mkv - max 500MB)",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "security": [{"ApiKeyAuth": []}],
                "parameters": [{
                    "name": "file",
                    "in": "formData",
                    "required": true,
                    "type": "file",
                    "description": "Video file"
                }],
                "responses": {
                    "200": {"description": "Video uploaded successfully"},
                    "400": {"description": "Invalid file type"},
                    "413": {"description": "File too large"}
                }
            }
        },
        "/files": {
            "get": {
                "tags": ["Files"],
                "summary": "List files",
                "description": "Get list of user's files with pagination, sorting and filtering",
                "produces": ["application/json"],
                "security": [{"ApiKeyAuth": []}],
                "parameters": [{
                    "name": "limit",
                    "in": "query",
                    "type": "integer",
                    "description": "Number of files (max 100)"
                }, {
                    "name": "offset",
                    "in": "query",
                    "type": "integer",
                    "description": "Number of files to skip"
                }, {
                    "name": "page",
                    "in": "query",
                    "type": "integer",
                    "description": "Page number"
                }, {
                    "name": "sort",
                    "in": "query",
                    "type": "string",
                    "description": "Sort by: date, name, size"
                }, {
                    "name": "type",
                    "in": "query",
                    "type": "string",
                    "description": "Filter by type: image, video, document"
                }],
                "responses": {
                    "200": {"description": "Files list"}
                }
            }
        },
        "/files/search": {
            "get": {
                "tags": ["Files"],
                "summary": "Search files",
                "description": "Search files by name",
                "produces": ["application/json"],
                "security": [{"ApiKeyAuth": []}],
                "parameters": [{
                    "name": "q",
                    "in": "query",
                    "type": "string",
                    "required": true,
                    "description": "Search query"
                }, {
                    "name": "limit",
                    "in": "query",
                    "type": "integer",
                    "description": "Number of results"
                }],
                "responses": {
                    "200": {"description": "Search results"}
                }
            }
        },
        "/files/stats": {
            "get": {
                "tags": ["Files"],
                "summary": "Get storage stats",
                "description": "Get storage statistics (total files, size breakdown by type)",
                "produces": ["application/json"],
                "security": [{"ApiKeyAuth": []}],
                "responses": {
                    "200": {"description": "Storage statistics"}
                }
            }
        },
        "/files/{id}": {
            "get": {
                "tags": ["Files"],
                "summary": "Get file by ID",
                "description": "Get file details including CloudFront URL",
                "produces": ["application/json"],
                "security": [{"ApiKeyAuth": []}],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "type": "string",
                    "required": true,
                    "description": "File ID"
                }],
                "responses": {
                    "200": {"description": "File details"},
                    "404": {"description": "File not found"}
                }
            },
            "delete": {
                "tags": ["Files"],
                "summary": "Delete file",
                "description": "Delete a single file from S3 and database",
                "produces": ["application/json"],
                "security": [{"ApiKeyAuth": []}],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "type": "string",
                    "required": true,
                    "description": "File ID"
                }],
                "responses": {
                    "204": {"description": "File deleted"},
                    "404": {"description": "File not found"}
                }
            }
        },
        "/files/delete": {
            "post": {
                "tags": ["Files"],
                "summary": "Bulk delete files",
                "description": "Delete multiple files at once",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "security": [{"ApiKeyAuth": []}],
                "parameters": [{
                    "name": "ids",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "ids": {
                                "type": "array",
                                "items": {"type": "string"}
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {"description": "Files deleted"}
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "Type 'Bearer' followed by a space and JWT token."
        }
    }
}